+++ src/py/mat3ra/made/tools/analyze/lattice/helpers.py
@@ -9,18 +9,10 @@ def get_material_with_conventional_lattice(material: MaterialWithBuildMetadata) 
 
 
 def get_material_with_primitive_lattice(
-    material: MaterialWithBuildMetadata, return_original_if_not_reduced=False
+    material: MaterialWithBuildMetadata, return_original_if_not_reduced=False, keep_orientation=True
 ) -> MaterialWithBuildMetadata:
     analyzer = LatticeMaterialAnalyzer(material=material)
-    material_with_primitive_lattice = analyzer.material_with_primitive_lattice
-    original_number_of_atoms = material.basis.number_of_atoms
-    primitive_structure_number_of_atoms = material_with_primitive_lattice.basis.number_of_atoms
-    if original_number_of_atoms == primitive_structure_number_of_atoms:
-        # Not reduced, return original material if requested, to avoid unnecessary editions
-        if return_original_if_not_reduced:
-            return material
-    # Reduced, return the primitive structure
-    return material_with_primitive_lattice
+    return analyzer.get_material_with_primitive_lattice_standard(return_original_if_not_reduced, keep_orientation)
 
 
 def get_lattice_type(material: MaterialWithBuildMetadata, tolerance=0.2, angle_tolerance=5) -> LatticeTypeEnum:
